% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frase.motivacional.R
\name{frase.motivacional}
\alias{frase.motivacional}
\title{Frase motivacional.}
\usage{
frase.motivacional()
}
\value{
Una frase motivacional.
}
\description{
Función que devuelve una frase motivacional.
}
\details{
Devuelve una frase escogida aleatoriamente entre un conjunto de frases obtenidas de \url{https://www.psicoactiva.com/blog/las-50-mejores-frases-de-motivacion-para-el-exito/}.
}
\author{
Carlos D. Luján
}
