% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.dos.numeros.R
\name{resta.dos.numeros}
\alias{resta.dos.numeros}
\title{Resta dos números}
\usage{
resta.dos.numeros(a, b)
}
\arguments{
\item{a}{minuendo.}

\item{b}{sustraendo.}
}
\value{
La resta de los dos números.
}
\description{
Función que resta dos números.
}
\examples{
resta.dos.numeros(2, 3)

x <- 2
y <- 3
resta.dos.numeros(x, y)

}
\author{
Carlos D. Luján
}
