% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma dos números}
\usage{
suma.dos.numeros(a, b)
}
\arguments{
\item{a}{primer sumando.}

\item{b}{segundo sumando.}
}
\value{
La suma de los dos números.
}
\description{
Función que suma dos números.
}
\examples{
suma.dos.numeros(2, 3)

x <- 2
y <- 3
suma.dos.numeros(x, y)

}
\author{
Carlos D. Luján
}
